Reorganize Data.Macaw.Memory modules
